test: improve fixture coverage and extract game data
- Extract game data from conftest.py into game_fixtures.py
- Add missing event types: RED_CARD, PENALTY_MISSED, BENCHED, OWN_GOAL
- Add missing assist type: CORNER_ASSIST
- Ensure >= 2 occurrences of each event type, goal type, and game result
- Simplify CI to just `uv run pytest`
- Rename misleading test name
- Add tests for own goals, penalty missed, benched, corner assists
