+++ .github/workflows/tests.yaml
@@ -18,4 +18,4 @@ jobs:
       - name: Install dependencies
         run: uv sync
       - name: Run tests
-        run: uv run pytest packages/maccabipediabot/tests packages/maccabipedia-mcp/tests packages/maccabistats/tests -v
+        run: uv run pytest -v
